+++ src/util/XDRStream.h
@@ -23,10 +23,10 @@ class XDRInputFileStream
 {
     std::ifstream mIn;
     std::vector<char> mBuf;
-    int mSizeLimit;
+    uint64_t mSizeLimit;
 
   public:
-    XDRInputFileStream(int sizeLimit = 0) : mSizeLimit{sizeLimit}
+    XDRInputFileStream(uint64_t sizeLimit = 0) : mSizeLimit{sizeLimit}
     {
     }
